+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `clusterHighlights` prop to the product
+
 ## [1.23.1] - 2020-11-05
 
 ### Changed